refactor: sepatae api logic
