+++ src/pages/login.tsx
@@ -4,7 +4,7 @@ import { useState, ReactElement } from "react";
 import HomeLayout from "@/containers/home/HomeLayout";
 import Login from "@/containers/home/Login";
 import Loading from "@/components/hint/Loading";
-import apiNext from "./api/apiNext";
+import apiNext from "../common/fetch/apiNext";
 
 import { useDispatch } from "react-redux";
 import { setUserInfo } from "@/common/redux/userInfoSlice";
