+++ src/pages/signup.tsx
@@ -7,7 +7,7 @@ import { setUserInfo } from "@/common/redux/userInfoSlice";
 import HomeLayout from "@/containers/home/HomeLayout";
 import SignUp from "@/containers/home/SignUp";
 import Loading from "@/components/hint/Loading";
-import apiNext from "./api/apiNext";
+import apiNext from "../common/fetch/apiNext";
 
 import type { ReactElement } from "react";
 import type { NextPageWithLayout } from "../pages/_app";
